Refactor: Moved gym hub tabs into its own component
